requirements.txt
pycryptodome
source venv/bin/activate
